Refactor SinglyLinkedList implementation file
Changed asserts from assert(pointer) to assert(pointer != nullptr) which is
much more clear.
Changed the parameter name 'data' to 'item' in insertion functions for consistency
with the header file.
Changed the parameter name 'source' to 'rhs' in both operator='s for consistency with
the header file.
Inlined some functions.